% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg RC class
The class linreg is a class of type RC that performs a linear regression given a
data frame and a formula. The class has several methods. print(), summary(), coef(), resid(), plot()
and pred().}
\value{
an object of class linreg
}
\description{
linreg RC class
The class linreg is a class of type RC that performs a linear regression given a
data frame and a formula. The class has several methods. print(), summary(), coef(), resid(), plot()
and pred().
}
\section{Fields}{

\describe{
\item{\code{X}}{matrix.}

\item{\code{y}}{numeric.}

\item{\code{b_hat}}{matrix.}

\item{\code{y_hat}}{numeric.}

\item{\code{e_hat}}{numeric.}

\item{\code{n}}{numeric.}

\item{\code{p}}{numeric.}

\item{\code{df}}{numeric.}

\item{\code{v_hat}}{numeric.}

\item{\code{v_b_hat_hat}}{matrix.}

\item{\code{t_values}}{numeric.}

\item{\code{p_values}}{numeric.}

\item{\code{data}}{data.frame.}

\item{\code{formula}}{formula.}

\item{\code{d}}{character.}
}}


\examples{
data(iris)
lin <- linreg$new(Petal.Length~Species, iris)
}
